Make `updateGroupMetrics` use precomputed contract metrics
